+++ scaleph-ui-react/src/pages/Project/Workspace/DataIntegration/SeaTunnel/Dag/components/node/steps/sink/sink-ftp-file-step.tsx
@@ -9,47 +9,47 @@ import DataSourceItem from "../dataSource";
 import FileSinkItem from "../common/file/fileSink";
 
 const SinkFtpFileStepForm: React.FC<ModalFormProps<Node>> = ({data, visible, onVisibleChange, onOK}) => {
-    const intl = getIntl(getLocale());
-    const [form] = Form.useForm();
+  const intl = getIntl(getLocale());
+  const [form] = Form.useForm();
 
-    useEffect(() => {
-        form.setFieldValue(STEP_ATTR_TYPE.stepTitle, data.data.label);
-    }, []);
+  useEffect(() => {
+    form.setFieldValue(STEP_ATTR_TYPE.stepTitle, data.data.label);
+  }, []);
 
-    return (
-        <XFlow>
-            <DrawerForm
-                title={data.data.label}
-                form={form}
-                initialValues={data.data.attrs}
-                open={visible}
-                onOpenChange={onVisibleChange}
-                grid={true}
-                width={780}
-                drawerProps={{
-                    styles: {body: {overflowY: 'scroll'}},
-                    closeIcon: null,
-                    destroyOnClose: true
-                }}
-                onFinish={(values) => {
-                    if (onOK) {
-                        onOK(values)
-                        return Promise.resolve(true)
-                    }
-                    return Promise.resolve(false)
-                }}
-            >
-                <ProFormText
-                    name={STEP_ATTR_TYPE.stepTitle}
-                    label={intl.formatMessage({id: 'pages.project.di.step.stepTitle'})}
-                    rules={[{required: true}, {max: 120}]}
-                    colProps={{span: 24}}
-                />
-                <DataSourceItem dataSource={'Ftp'}/>
-                <FileSinkItem/>
-            </DrawerForm>
-        </XFlow>
-    );
+  return (
+    <XFlow>
+      <DrawerForm
+        title={data.data.label}
+        form={form}
+        initialValues={data.data.attrs}
+        open={visible}
+        onOpenChange={onVisibleChange}
+        grid={true}
+        width={780}
+        drawerProps={{
+          styles: {body: {overflowY: 'scroll'}},
+          closeIcon: null,
+          destroyOnClose: true
+        }}
+        onFinish={(values) => {
+          if (onOK) {
+            onOK(values)
+            return Promise.resolve(true)
+          }
+          return Promise.resolve(false)
+        }}
+      >
+        <ProFormText
+          name={STEP_ATTR_TYPE.stepTitle}
+          label={intl.formatMessage({id: 'pages.project.di.step.stepTitle'})}
+          rules={[{required: true}, {max: 120}]}
+          colProps={{span: 24}}
+        />
+        <DataSourceItem dataSource={'Ftp'}/>
+        <FileSinkItem/>
+      </DrawerForm>
+    </XFlow>
+  );
 };
 
 export default SinkFtpFileStepForm;
